requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
boto3==1.34.0
stripe==7.11.0
sendgrid==6.11.0
pillow==10.2.0
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
psutil==5.9.6
matplotlib==3.8.2
scipy==1.11.4
reportlab==4.0.8
qrcode[pil]==7.4.2
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
alembic==1.13.1
